Fixed Flight Controller (Discounts added, format fixed). Added a get Table for seating and discounts table added on Add Airlines

+++ HotelBackend/src/main/java/com/dailycodebuffer/Model/Seating.java
@@ -14,9 +14,9 @@ public class Seating {
     private Integer numberOfRows;
     private Integer startRow;
     private Integer endRow;
-    private Integer aislePrice;
-    private Integer middlePrice;
-    private Integer windowPrice;
+    private Long aislePrice;
+    private Long middlePrice;
+    private Long windowPrice;
     private Boolean extraLegroom;
 
     @ElementCollection
